trying to make solution fully 2.1 compatible
maybe need this to get functional tests working
For all I know it may have already been fully compatible with 2.1
However there was a hint somewhere that we should be using
Microsoft.AspNetCore.App rather than Microsoft.AspNetCore.All
as the main package.  Other changes followed from that.
Why all this nonsense when the SDk was correctly set to Microsoft.AspNetCore.All?
why does this have to be duplicated with a package?