Refactor key handling in ActorService
This commit introduces several key changes to improve the management and utilization of cryptographic keys within the system. Firstly, the private key variable in `ActorService` has been refactored from `PrivateKey` to `_privateKey`, affecting its initialization and usage in the `SendSignedRequest` method. Additionally, a `publickeyid` parameter has been added to the `SendSignedRequest` method in both `ActorService` and the `IActorService` interface, necessitating updates across all method calls to include this new argument.

The approach to key management has been further streamlined by removing the code responsible for reading public and private keys from files within the `SendSignedRequest` method. Now, the private key is directly set from the `_privateKey` field. Correspondingly, the logging within this method has been updated to log the `publickeyid` instead of the public key, aligning with the new parameter's introduction.

Moreover, the `ActivityPubManager` class has seen the addition of a `publickeyid` field, constructed using the `alias` field. This addition plays a crucial role in setting the `id` field of the `publicKey` object in the `GetActor` method and is essential for the updated calls to `SendSignedRequest`, ensuring the correct passage and utilization of the public key ID in the HTTP signature process.

These enhancements collectively fortify the system's key handling capabilities, offering a more explicit and flexible management of public and private keys.